+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [] - 2026-02-
+- Added factory interface for building a Ui5Context
+
 ## [4.1.5] - 2026-02-21
 - Added `Platform` value to enum `ReadScope`
 - Changed name of enum `ReadScope` to `Scope`